requirements.txt
# Image Processor — Dependencies
# Install with: pip install -r requirements.txt

fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pillow>=10.0.0
python-dotenv>=1.0.0
cloudinary>=1.36.0
